INTRD-10571 : add mising query

+++ opencell-admin/web/src/main/java/org/meveo/admin/action/billing/CreationInvoiceBean.java
@@ -356,6 +356,9 @@ public class CreationInvoiceBean extends CustomFieldBean<Invoice> {
         seller=sellerService.refreshOrRetrieve(seller);
         TaxInfo taxInfo = taxMappingService.determineTax(selectedCharge, seller, ua, usageDate);
 
+        if(entity.getId() == null) {
+            invoiceService.create(entity);
+        }
         // AKK check what happens with tax
         RatedTransaction ratedTransaction = new RatedTransaction(usageDate, unitAmountWithoutTax, unitAmountWithTax, null, quantity, null, null, null, RatedTransactionStatusEnum.BILLED, ua.getWallet(),
             ua.getBillingAccount(), ua, selectInvoiceSubCat, parameter1, parameter2, parameter3, null, orderNumber, null, null, null, null, null, null, selectedCharge.getCode(), description, rtStartDate, rtEndDate,
@@ -667,6 +670,9 @@ public class CreationInvoiceBean extends CustomFieldBean<Invoice> {
     			entity = invoiceService.retrieveIfNotManaged(entity);
     		}
     	}
+        if(entity.getId() == null) {
+            invoiceService.create(entity);
+        }
         if (!ratedTransactionsToRemove.isEmpty()) {
             ratedTransactionsToSave.removeAll(ratedTransactionsToRemove);
             ratedTransactionService.cancelRatedTransactions(ratedTransactionsToRemove
@@ -675,7 +681,9 @@ public class CreationInvoiceBean extends CustomFieldBean<Invoice> {
                     .collect(Collectors.toList()));
         }
         detachRts();
-        invoiceAgregateService.deleteInvoiceAggregates(entity.getId());
+        if(entity.getId() != null) {
+            invoiceAgregateService.deleteInvoiceAggregates(entity.getId());
+        }
         Map<String, CategoryInvoiceAgregate> categoryInvoiceAggregateMap = new HashMap<>();
         List<SubCategoryInvoiceAgregate> newSubCategoryInvoiceAggregates = new ArrayList<>();
         for (RatedTransaction ratedTransaction : ratedTransactionsToSave) {
